refactor(generic-content-plugin): Rename from `root-with-optionals` to `type-generic-content`
