--- packages/editor/src/plugins/root-with-optionals/editor.tsx
@@ -1,6 +1,0 @@
-import type { RootWithOptionalsProps } from '.'
-import { RootWithOptionalsRenderer } from './renderer'
-
-export function RootWithOptionalsEditor({ state }: RootWithOptionalsProps) {
-  return <RootWithOptionalsRenderer content={state.content.render()} />
-}
