--- packages/editor/src/plugins/root-with-optionals/renderer.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react'
-
-export function RootWithOptionalsRenderer({ content }: { content: ReactNode }) {
-  return <>{content}</>
-}
