+++ packages/editor/src/plugins/rows/editor.tsx
@@ -13,9 +13,7 @@ export function RowsEditor({ state, config, id }: RowsProps) {
   const parentType = selectParentPluginType(store.getState(), id)
   const isParentTemplatePlugin = parentType?.startsWith('type-')
   const showLargeAddButton =
-    parentType === EditorPluginType.Article ||
-    parentType === EditorPluginType.RootWithOptionals ||
-    isParentTemplatePlugin
+    parentType === EditorPluginType.Article || isParentTemplatePlugin
 
   function insertRowWithSuggestionsOpen(insertIndex: number) {
     const textPluginWithSuggestions = {
