+++ packages/editor/src/plugins/serlo-template-plugins/generic-content/editor.tsx
@@ -1,0 +1,6 @@
+import type { GenericContentTypePluginProps } from '.'
+import { GenericContentTypeRenderer } from './renderer'
+
+export function GenericContentTypeEditor({ state }: GenericContentTypePluginProps) {
+  return <GenericContentTypeRenderer content={state.content.render()} />
+}
