+++ packages/editor/src/plugins/serlo-template-plugins/generic-content/renderer.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export function GenericContentTypeRenderer({ content }: { content: ReactNode }) {
+  return <>{content}</>
+}
