+++ packages/editor/src/types/editor-plugins.ts
@@ -16,11 +16,11 @@ import type { MultimediaPluginState } from '@editor/plugins/multimedia'
 import { PageLayoutPluginState } from '@editor/plugins/page-layout'
 import { PagePartnersPluginState } from '@editor/plugins/page-partners'
 import { PageTeamPluginState } from '@editor/plugins/page-team'
-import { RootWithOptionalsState } from '@editor/plugins/root-with-optionals'
 import type { RowsPluginState } from '@editor/plugins/rows'
 import type { ScMcExercisePluginState } from '@editor/plugins/sc-mc-exercise'
 import { SerloTablePluginState } from '@editor/plugins/serlo-table'
 import { TextExerciseGroupTypePluginState } from '@editor/plugins/serlo-template-plugins/exercise-group/text-exercise-group'
+import { GenericContentTypePluginState } from '@editor/plugins/serlo-template-plugins/generic-content'
 import type { SolutionPluginState } from '@editor/plugins/solution'
 import { SpoilerPluginState } from '@editor/plugins/spoiler'
 import type {
@@ -49,11 +49,6 @@ export interface EditorAnchorDocument {
 export interface EditorArticleDocument {
   plugin: EditorPluginType.Article
   state: PrettyStaticState<ArticlePluginState>
-  id?: string
-}
-export interface EditorRootWithOptionalsDocument {
-  plugin: EditorPluginType.RootWithOptionals
-  state: PrettyStaticState<RootWithOptionalsState>
   id?: string
 }
 export interface EditorArticleIntroductionDocument {
@@ -222,6 +217,11 @@ export interface EditorTemplateExerciseGroupDocument {
     unrevisedRevisions?: number
     license?: License
   }
+}
+export interface EditorTemplateGenericContentDocument {
+  plugin: TemplatePluginType.GenericContent
+  state: PrettyStaticState<GenericContentTypePluginState>
+  id?: string
 }
 
 export type SupportedEditorDocument =
